backend/requirements.txt
fastapi>=0.115.6
uvicorn[standard]>=0.34.0
pydantic>=2.12.0
motor>=3.7.0
python-dotenv>=1.0.1
openai>=1.59.7
